OBPIH-6974 Move to resolve tab
